pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "position-tracker"
version = "2.0.0"
description = "SaaS SERP position tracker with PostgreSQL + ClickHouse"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
    # Database - PostgreSQL
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # Database - ClickHouse
    "clickhouse-connect>=0.8",
    # Background tasks
    "celery[redis]>=5.4",
    "redis>=5.0",
    # Auth
    "pyjwt>=2.9",
    "argon2-cffi>=23.1",
    # Config & validation
    "pydantic-settings>=2.5",
    "email-validator>=2.1",
    # Alembic sync driver
    "psycopg2-binary>=2.9",
    # HTTP client (for collectors)
    "httpx>=0.27",
    # Encryption (tenant credentials)
    "cryptography>=43.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Monitoring
    "prometheus-client>=0.21",
    "sentry-sdk[fastapi]>=2.0",
    # Entity clustering (ProLEA + TaxoAdapt pipeline)
    "numpy>=1.26",
    "scipy>=1.12",
    "pgvector>=0.3",
    # HTML parsing (GEO audit)
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "defusedxml>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pytest-cov>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
